# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridLine AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/runs/{run_id}/approval": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Decide Run Approval",
        "description": "Record the decision and resume the graph; progress arrives as ``agent.step`` events.",
        "operationId": "decideRunApproval",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDecision"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRun"
                }
              }
            }
          },
          "404": {
            "description": "No such run"
          },
          "409": {
            "description": "The run is not waiting for approval"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowAction": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citation Ids"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "candidate_id",
          "tool",
          "input",
          "label",
          "rationale",
          "citation_ids",
          "requires_approval"
        ],
        "title": "WorkflowAction"
      },
      "WorkflowActionResult": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "status": {
            "type": "string",
            "enum": [
              "executed",
              "unchanged",
              "rejected",
              "failed"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "affected_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Entities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "tool",
          "status",
          "message",
          "affected_entities"
        ],
        "title": "WorkflowActionResult"
      },
      "WorkflowActionVerification": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "failed"
            ],
            "title": "Status"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/WorkflowCheck"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "tool",
          "status",
          "checks"
        ],
        "title": "WorkflowActionVerification"
      },
      "WorkflowApprovalOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "approval_gate",
            "title": "Node",
            "default": "approval_gate"
          },
          "approval_id": {
            "type": "string",
            "title": "Approval Id"
          },
          "action_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Ids"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "approve",
                  "reject"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "approval_id",
          "action_ids",
          "decision",
          "note",
          "decided_at"
        ],
        "title": "WorkflowApprovalOutput"
      },
      "WorkflowAssessmentOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "assess",
            "title": "Node",
            "default": "assess"
          },
          "hazard": {
            "type": "string",
            "enum": [
              "landslide",
              "flood"
            ],
            "title": "Hazard"
          },
          "band": {
            "type": "string",
            "enum": [
              "normal",
              "watch",
              "warning",
              "critical"
            ],
            "title": "Band"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "grounded": {
            "type": "boolean",
            "title": "Grounded"
          },
          "cited_count": {
            "type": "integer",
            "title": "Cited Count"
          },
          "ungrounded_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ungrounded Ids"
          },
          "affected_zone_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Zone Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "hazard",
          "band",
          "confidence",
          "grounded",
          "cited_count",
          "ungrounded_ids",
          "affected_zone_ids"
        ],
        "title": "WorkflowAssessmentOutput"
      },
      "WorkflowCheck": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "expected": {
            "type": "string",
            "title": "Expected"
          },
          "observed": {
            "type": "string",
            "title": "Observed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "passed",
          "expected",
          "observed"
        ],
        "title": "WorkflowCheck"
      },
      "WorkflowChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "document_title",
          "section",
          "kind",
          "similarity"
        ],
        "title": "WorkflowChunk"
      },
      "WorkflowClaim": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citation Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "citation_ids"
        ],
        "title": "WorkflowClaim"
      },
      "WorkflowCompletionOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "complete",
            "title": "Node",
            "default": "complete"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "completed",
              "completed_with_failures",
              "rejected"
            ],
            "title": "Outcome"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEntityState"
            },
            "type": "array",
            "title": "Entities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "outcome",
          "entities"
        ],
        "title": "WorkflowCompletionOutput"
      },
      "WorkflowDecision": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "WorkflowDecision"
      },
      "WorkflowEntity": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "table",
          "id",
          "name"
        ],
        "title": "WorkflowEntity"
      },
      "WorkflowEntityState": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "name",
          "fields"
        ],
        "title": "WorkflowEntityState"
      },
      "WorkflowEvidenceOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "retrieve",
            "title": "Node",
            "default": "retrieve"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/WorkflowChunk"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "queries",
          "chunks"
        ],
        "title": "WorkflowEvidenceOutput"
      },
      "WorkflowExecutionOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "execute",
            "title": "Node",
            "default": "execute"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/WorkflowActionResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "results"
        ],
        "title": "WorkflowExecutionOutput"
      },
      "WorkflowGraphOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "query_graph",
            "title": "Node",
            "default": "query_graph"
          },
          "start": {
            "$ref": "#/components/schemas/WorkflowEntity"
          },
          "entity_count": {
            "type": "integer",
            "title": "Entity Count"
          },
          "edge_count": {
            "type": "integer",
            "title": "Edge Count"
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/WorkflowPath"
            },
            "type": "array",
            "title": "Paths"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "start",
          "entity_count",
          "edge_count",
          "paths"
        ],
        "title": "WorkflowGraphOutput"
      },
      "WorkflowObserveOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "observe",
            "title": "Node",
            "default": "observe"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSignal"
            },
            "type": "array",
            "title": "Signals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "headline",
          "signals"
        ],
        "title": "WorkflowObserveOutput"
      },
      "WorkflowPath": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEntity"
            },
            "type": "array",
            "title": "Nodes"
          },
          "relations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relations"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citation Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "relations",
          "citation_ids"
        ],
        "title": "WorkflowPath",
        "description": "``nodes[i] --relations[i]--> nodes[i+1]``, each hop backed by the KG edge in ``citation_ids[i]``."
      },
      "WorkflowPlanOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "recommend",
            "title": "Node",
            "default": "recommend"
          },
          "provider": {
            "type": "string",
            "enum": [
              "ollama",
              "mock"
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/WorkflowAction"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "provider",
          "model",
          "fallback_reason",
          "duration_ms",
          "candidate_count",
          "actions"
        ],
        "title": "WorkflowPlanOutput"
      },
      "WorkflowReasoningOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "reason",
            "title": "Node",
            "default": "reason"
          },
          "provider": {
            "type": "string",
            "enum": [
              "ollama",
              "mock"
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/WorkflowClaim"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "provider",
          "model",
          "fallback_reason",
          "duration_ms",
          "summary",
          "claims"
        ],
        "title": "WorkflowReasoningOutput"
      },
      "WorkflowReceiveOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "receive",
            "title": "Node",
            "default": "receive"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "hazard": {
            "type": "string",
            "enum": [
              "landslide",
              "flood"
            ],
            "title": "Hazard"
          },
          "failure_kind": {
            "type": "string",
            "title": "Failure Kind"
          },
          "asset": {
            "$ref": "#/components/schemas/WorkflowEntity"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "zone_name": {
            "type": "string",
            "title": "Zone Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sim_time": {
            "type": "string",
            "format": "date-time",
            "title": "Sim Time"
          },
          "citation_id": {
            "type": "string",
            "title": "Citation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "event_id",
          "hazard",
          "failure_kind",
          "asset",
          "zone_id",
          "zone_name",
          "description",
          "sim_time",
          "citation_id"
        ],
        "title": "WorkflowReceiveOutput"
      },
      "WorkflowRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "trigger_event_id": {
            "type": "string",
            "title": "Trigger Event Id"
          },
          "hazard": {
            "type": "string",
            "enum": [
              "landslide",
              "flood"
            ],
            "title": "Hazard"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "trigger_event_id",
          "hazard",
          "zone_id",
          "asset_id",
          "provider",
          "model",
          "started_at",
          "steps"
        ],
        "title": "WorkflowRun"
      },
      "WorkflowSignal": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "event_type",
          "source",
          "zone_id",
          "severity",
          "summary"
        ],
        "title": "WorkflowSignal"
      },
      "WorkflowStep": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "node": {
            "type": "string",
            "enum": [
              "receive",
              "observe",
              "query_graph",
              "retrieve",
              "reason",
              "assess",
              "recommend",
              "approval_gate",
              "execute",
              "verify",
              "complete"
            ],
            "title": "Node"
          },
          "index": {
            "type": "integer",
            "maximum": 11.0,
            "minimum": 1.0,
            "title": "Index"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "done",
              "waiting",
              "failed"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "output": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/WorkflowReceiveOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowObserveOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowGraphOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowEvidenceOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowReasoningOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowAssessmentOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowPlanOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowApprovalOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowExecutionOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowVerificationOutput"
                  },
                  {
                    "$ref": "#/components/schemas/WorkflowCompletionOutput"
                  }
                ],
                "discriminator": {
                  "propertyName": "node",
                  "mapping": {
                    "approval_gate": "#/components/schemas/WorkflowApprovalOutput",
                    "assess": "#/components/schemas/WorkflowAssessmentOutput",
                    "complete": "#/components/schemas/WorkflowCompletionOutput",
                    "execute": "#/components/schemas/WorkflowExecutionOutput",
                    "observe": "#/components/schemas/WorkflowObserveOutput",
                    "query_graph": "#/components/schemas/WorkflowGraphOutput",
                    "reason": "#/components/schemas/WorkflowReasoningOutput",
                    "receive": "#/components/schemas/WorkflowReceiveOutput",
                    "recommend": "#/components/schemas/WorkflowPlanOutput",
                    "retrieve": "#/components/schemas/WorkflowEvidenceOutput",
                    "verify": "#/components/schemas/WorkflowVerificationOutput"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "node",
          "index",
          "status",
          "started_at",
          "finished_at",
          "duration_ms",
          "output",
          "error"
        ],
        "title": "WorkflowStep",
        "description": "The ``agent.step`` payload: the latest state of one node of one run."
      },
      "WorkflowVerificationOutput": {
        "properties": {
          "node": {
            "type": "string",
            "const": "verify",
            "title": "Node",
            "default": "verify"
          },
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "partially_verified",
              "failed"
            ],
            "title": "Status"
          },
          "per_action": {
            "items": {
              "$ref": "#/components/schemas/WorkflowActionVerification"
            },
            "type": "array",
            "title": "Per Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node",
          "status",
          "per_action"
        ],
        "title": "WorkflowVerificationOutput"
      },
...
    }
  }
}